CameraFragment 화면 및 기능 구현 (#22)
* refactor: 클래스 명 변경

* refactor: object로 util함수 분리

* refactor: 리뷰 반영@EvergreenTree97, @ham2174

* feat: CameraActivity 테마 적용

* refactor: 필요없는 import 제거

* refactor: 시작화면 CameraActivity 로 변경

* refactor: cameraActivity 화면 재구성

* refactor: ratio 초기화하는 방법 재설정

* feat: CameraFragment, CameraResultFragment 생성
