+++ presentation/src/main/java/team/jsv/icec/ui/camera/CameraViewModel.kt
@@ -1,19 +1,30 @@
 package team.jsv.icec.ui.camera
 
+import android.graphics.Bitmap
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import team.jsv.icec.base.BaseViewModel
 
 
 class CameraViewModel : BaseViewModel() {
-    private val _ratioState = MutableLiveData<Int>()
+    private val _ratioState = MutableLiveData<Int>().apply { value = SettingRatio.RATIO_1_1.id }
     val ratioState: LiveData<Int> get() = _ratioState
 
-    fun initSetRatio() {
-        _ratioState.value = SettingRatio.RATIO_1_1.id
-    }
+    private val _bitmapImage = MutableLiveData<Bitmap>()
+    val bitmapImage: LiveData<Bitmap> get() = _bitmapImage
+
+    private val _isFrontCamera = MutableLiveData<Boolean>().apply { value = true }
+    val isFrontCamera: LiveData<Boolean> get() = _isFrontCamera
 
     fun setRatioState(value: Int) {
         _ratioState.value = value
+    }
+
+    fun setBitmapImage(image: Bitmap) {
+        _bitmapImage.value = image
+    }
+
+    fun setIsFrontCamera(value : Boolean) {
+        _isFrontCamera.value = value
     }
 }
