+++ presentation/src/main/res/navigation/camera_nav_graph.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/camera_nav_graph"
+    app:startDestination="@id/cameraFragment">
+
+    <fragment
+        android:id="@+id/cameraFragment"
+        android:name="team.jsv.icec.ui.camera.CameraFragment"
+        android:label="CameraFragment">
+        <action
+            android:id="@+id/action_cameraFragment_to_cameraResultFragment"
+            app:destination="@id/cameraResultFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/cameraResultFragment"
+        android:name="team.jsv.icec.ui.camera.CameraResultFragment"
+        android:label="CameraResultFragment" />
+</navigation>
